requirements.txt
Django==5.0.2
django-filter==23.5
Flask==3.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
matplotlib==3.8.3
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.0
pillow==10.2.0
pytchat==0.5.5
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
youtube-search-python==1.6.6
